@@ -24,45 +24,45 @@
     "codeclimate": "codeclimate-test-reporter < coverage/lcov.info"
   },
   "dependencies": {
-    "@types/lodash": "^4.14.63",
+    "@types/lodash": "^4.14.116",
+    "@types/reach__router": "^1.0.1",
     "@types/react": "^16.0.20",
     "@types/react-dom": "^16.0.2",
     "@types/react-helmet": "5.0.0",
-    "@types/react-redux": "^5.0.9",
-    "@types/react-router": "^4.0.14",
-    "@types/react-router-dom": "^4.0.8",
+    "@types/react-redux": "^5.0.21",
+    "@types/semantic-ui": "^2.2.6",
     "babel-eslint": "^7.2.1",
     "change-case": "^3.0.1",
     "codeclimate-test-reporter": "^0.4.1",
     "disqus-react": "^1.0.5",
-    "gatsby": "latest",
-    "gatsby-link": "latest",
-    "gatsby-plugin-glamor": "latest",
-    "gatsby-plugin-google-analytics": "latest",
-    "gatsby-plugin-manifest": "latest",
-    "gatsby-plugin-offline": "latest",
-    "gatsby-plugin-sharp": "latest",
-    "gatsby-plugin-typescript": "latest",
-    "gatsby-remark-autolink-headers": "latest",
-    "gatsby-remark-copy-linked-files": "latest",
-    "gatsby-remark-images": "latest",
-    "gatsby-remark-prismjs": "latest",
-    "gatsby-source-filesystem": "latest",
-    "gatsby-transformer-json": "latest",
-    "gatsby-transformer-remark": "latest",
-    "gatsby-transformer-sharp": "latest",
+    "gatsby": "next",
+    "gatsby-link": "next",
+    "gatsby-plugin-glamor": "next",
+    "gatsby-plugin-google-analytics": "next",
+    "gatsby-plugin-manifest": "next",
+    "gatsby-plugin-offline": "next",
+    "gatsby-plugin-sharp": "next",
+    "gatsby-plugin-typescript": "^2.0.0-rc.4",
+    "gatsby-remark-autolink-headers": "next",
+    "gatsby-remark-copy-linked-files": "next",
+    "gatsby-remark-images": "next",
+    "gatsby-remark-prismjs": "next",
+    "gatsby-source-filesystem": "next",
+    "gatsby-transformer-json": "next",
+    "gatsby-transformer-remark": "next",
+    "gatsby-transformer-sharp": "next",
     "graphql-code-generator": "^0.5.2",
     "gray-matter": "^2.1.1",
     "lodash": "^4.17.4",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
+    "react": "^16.4.2",
+    "react-dom": "^16.4.2",
     "react-helmet": "5.0.3",
-    "react-redux": "^5.0.6",
+    "react-redux": "^5.0.7",
     "redux-devtools-extension": "^2.13.2",
-    "semantic-ui-react": "^0.74.2",
+    "semantic-ui-react": "^0.82.3",
     "slash": "1.0.0",
     "ts-loader": "^2.3.7",
-    "typescript": "2.5.3"
+    "typescript": "^2.8.0"
   },
   "devDependencies": {
     "@storybook/addon-actions": "^3.2.14",
