@@ -6,32 +6,14 @@ import Helmet from "react-helmet";
 
 const config = require("../gatsby-config.js");
 
-// Load production style
-let styles: string;
-if (process.env.NODE_ENV === `production`) {
-  try {
-    styles = require("!raw-loader!../public/styles.css");
-  } catch (err) {
-    console.log(err);
-  }
-}
-
 interface HtmlProps {
   body: any;
   postBodyComponents: any;
   headComponents: any;
 }
 
-// Use `module.exports` to be compliante with `webpack-require` import method
-module.exports = (props: HtmlProps) => {
+export default (props: HtmlProps) => {
   const head = Helmet.rewind();
-
-  const css = (process.env.NODE_ENV === `production`) ?
-    <style
-      id="gatsby-inlined-css"
-      dangerouslySetInnerHTML={{ __html: styles }}
-    />
-    : null;
 
   const verification = config.siteMetadata && config.siteMetadata.googleVerification ? <meta
     name="google-site-verification"
@@ -51,7 +33,6 @@ module.exports = (props: HtmlProps) => {
         {head.title.toComponent()}
         {head.meta.toComponent()}
         {head.link.toComponent()}
-        {css}
         {verification}
       </head>
       <body>
