@@ -1,11 +1,9 @@
 import * as React from "react";
-import { Router } from "react-router-dom";
 import { Provider } from "react-redux";
 
 import { store } from "./src/store";
 
-exports.replaceRouterComponent = ({ history }) =>
-  ({ children }) =>
+export const wrapRootElement = ({ element }) =>
     <Provider store={store} >
-      <Router history={history}>{children}</Router>
+      {element}
     </Provider>;
