@@ -12,5 +12,5 @@ export interface MenuItem {
 export interface MenuProps extends React.HTMLProps<HTMLDivElement> {
   items: MenuItem[];
   pathname: string;
-  Link: React.ComponentClass<GatsbyLinkProps> | any;
+  Link: React.ComponentClass<GatsbyLinkProps<any>> | any;
 }
