@@ -5,11 +5,12 @@ import { GatsbyLinkProps } from "gatsby-link";
 import { StoreState } from "../../store";
 import { MenuProps, MenuItem } from "../Menu";
 import { Menu, Icon, Sidebar } from "semantic-ui-react";
+import { SemanticICONS } from "semantic-ui-react";
 
 interface SidebarMenuProps extends MenuProps {
   visible?: boolean;
   dispatch?: Dispatch<any>;
-  Link: React.ComponentClass<GatsbyLinkProps>;
+  Link: React.ComponentClass<GatsbyLinkProps<any>>;
 }
 
 export const SidebarMenu = ({ items, pathname, Link, visible }: SidebarMenuProps) => {
@@ -22,7 +23,7 @@ export const SidebarMenu = ({ items, pathname, Link, visible }: SidebarMenuProps
         const active = isActive(item);
         return (
           <Menu.Item as={Link} to={item.path} active={active} key={item.path}>
-            <Icon name={item.icon} />
+            <Icon name={item.icon as SemanticICONS} />
             {item.name}
           </Menu.Item>
         );
