@@ -1,26 +1,26 @@
 import * as React from "react";
-import Link from "gatsby-link";
+import { Link } from "gatsby";
+import { StaticQuery, graphql } from "gatsby";
 import { Header, Grid, Card, List, Container, Feed, Segment, Comment } from "semantic-ui-react";
 import { MarkdownRemarkConnection, ImageSharp } from "../graphql-types";
 import BlogTitle from "../components/BlogTitle";
 import TagsCard from "../components/TagsCard/TagsCard";
 import BlogPagination from "../components/BlogPagination/BlogPagination";
 import { get } from "lodash";
+import {withLayout, LayoutProps} from "../components/Layout";
+import { MarkdownRemark } from "../graphql-types";
 
-interface BlogProps {
+interface BlogProps extends LayoutProps {
   data: {
     tags: MarkdownRemarkConnection;
     posts: MarkdownRemarkConnection;
   };
-  pathContext: {
+  pageContext: {
     tag?: string; // only set into `templates/tags-pages.tsx`
-  };
-  location: {
-    pathname: string;
   };
 }
 
-export default (props: BlogProps) => {
+const BlogPage = (props: BlogProps) => {
   const tags = props.data.tags.group;
   const posts = props.data.posts.edges;
   const { pathname } = props.location;
@@ -29,17 +29,17 @@ export default (props: BlogProps) => {
   // TODO export posts in a proper component
   const Posts = (
     <Container>
-      {posts.map(({ node }) => {
+      {posts.map(({ node }: {node: MarkdownRemark}) => {
         const { frontmatter, timeToRead, fields: { slug }, excerpt } = node;
         const avatar = frontmatter.author.avatar.children[0] as ImageSharp;
-        const cover = get(frontmatter, "image.children.0.responsiveResolution", {});
+        const cover = get(frontmatter, "image.children.0.fixed", {});
 
         const extra = (
           <Comment.Group>
             <Comment>
               <Comment.Avatar
-                src={avatar.responsiveResolution.src}
-                srcSet={avatar.responsiveResolution.srcSet}
+                src={avatar.fixed.src}
+                srcSet={avatar.fixed.srcSet}
               />
               <Comment.Content>
                 <Comment.Author style={{ fontWeight: 400 }}>
@@ -89,13 +89,15 @@ export default (props: BlogProps) => {
             </Segment>
           </div>
           <div>
-            <TagsCard Link={Link} tags={tags} tag={props.pathContext.tag} />
+            <TagsCard Link={Link} tags={tags} tag={props.pageContext.tag} />
           </div>
         </Grid>
       </Segment>
     </Container>
   );
 };
+
+export default withLayout(BlogPage);
 
 export const pageQuery = graphql`
 query PageBlog {
@@ -130,7 +132,7 @@ query PageBlog {
           image {
           	children {
               ... on ImageSharp {
-                responsiveResolution(width: 700, height: 100) {
+                fixed(width: 700, height: 100) {
                   src
                   srcSet
                 }
@@ -142,7 +144,7 @@ query PageBlog {
             avatar {
               children {
                 ... on ImageSharp {
-                  responsiveResolution(width: 35, height: 35) {
+                  fixed(width: 35, height: 35) {
                     src
                     srcSet
                   }
