@@ -1,5 +1,6 @@
 import * as React from "react";
 import Blog from "../pages/blog";
+import { graphql } from "gatsby";
 
 export default Blog;
 
@@ -39,7 +40,7 @@ query TemplateBlogPage($skip: Int) {
           image {
           	children {
               ... on ImageSharp {
-                responsiveResolution(width: 700, height: 100) {
+                fixed(width: 700, height: 100) {
                   src
                   srcSet
                 }
@@ -51,7 +52,7 @@ query TemplateBlogPage($skip: Int) {
             avatar {
               children {
                 ... on ImageSharp {
-                  responsiveResolution(width: 35, height: 35) {
+                  fixed(width: 35, height: 35) {
                     src
                     srcSet
                   }
