@@ -1,12 +1,14 @@
 import * as React from "react";
-import Link from "gatsby-link";
+import { Link } from "gatsby";
 import { get } from "lodash";
 import { Header, Container, Segment, Icon, Label, Button, Grid, Card, Image, Item, Comment } from "semantic-ui-react";
 import { MarkdownRemark, ImageSharp, MarkdownRemarkConnection, Site } from "../graphql-types";
 import BlogTitle from "../components/BlogTitle";
 import { DiscussionEmbed } from "disqus-react";
+import {withLayout, LayoutProps} from "../components/Layout";
+import { graphql } from "gatsby";
 
-interface BlogPostProps {
+interface BlogPostProps extends LayoutProps {
   data: {
     post: MarkdownRemark;
     recents: MarkdownRemarkConnection;
@@ -14,7 +16,7 @@ interface BlogPostProps {
   };
 }
 
-export default (props: BlogPostProps) => {
+const BlogPostPage = (props: BlogPostProps) => {
   const { frontmatter, html, timeToRead } = props.data.post;
   const avatar = frontmatter.author.avatar.children[0] as ImageSharp;
 
@@ -24,13 +26,13 @@ export default (props: BlogPostProps) => {
   const recents = props.data.recents.edges
     .map(({ node }) => {
       const recentAvatar = node.frontmatter.author.avatar.children[0] as ImageSharp;
-      const recentCover = get(node, "frontmatter.image.children.0.responsiveResolution", {});
+      const recentCover = get(node, "frontmatter.image.children.0.fixed", {});
       const extra = (
         <Comment.Group>
           <Comment>
             <Comment.Avatar
-              src={recentAvatar.responsiveResolution.src}
-              srcSet={recentAvatar.responsiveResolution.srcSet}
+              src={recentAvatar.fixed.src}
+              srcSet={recentAvatar.fixed.srcSet}
             />
             <Comment.Content>
               <Comment.Author style={{ fontWeight: 400 }}>
@@ -56,16 +58,17 @@ export default (props: BlogPostProps) => {
       );
     });
 
-  const cover = get(frontmatter, "image.children.0.responsiveResolution", {} );
+  const cover = get(frontmatter, "image.children.0.fixed", {} );
   return (
     <Container>
       <BlogTitle />
       <Segment vertical style={{ border: "none" }}>
         <Item.Group>
           <Item>
-            <Item.Image size="tiny" shape="circular"
-              src={avatar.responsiveResolution.src}
-              srcSet={avatar.responsiveResolution.srcSet}
+            <Item.Image size="tiny"
+              src={avatar.fixed.src}
+              srcSet={avatar.fixed.srcSet}
+              circular
             />
             <Item.Content>
               <Item.Description>{frontmatter.author.id}</Item.Description>
@@ -79,7 +82,7 @@ export default (props: BlogPostProps) => {
       <Image
         {...cover}
         fluid
-      />
+      /> 
       <Segment vertical
         style={{ border: "none" }}
         dangerouslySetInnerHTML={{
@@ -93,7 +96,7 @@ export default (props: BlogPostProps) => {
         && props.data.site.siteMetadata
         && props.data.site.siteMetadata.disqus
         && <Segment vertical>
-            <DiscussionEmbed shortname={props.data.site.siteMetadata.disqus}/>
+            <DiscussionEmbed shortname={props.data.site.siteMetadata.disqus} config={{}}/>
         </Segment>
       }
       <Segment vertical>
@@ -104,6 +107,8 @@ export default (props: BlogPostProps) => {
     </Container>
   );
 };
+
+export default withLayout(BlogPostPage);
 
 export const pageQuery = graphql`
   query TemplateBlogPost($slug: String!) {
@@ -128,7 +133,7 @@ export const pageQuery = graphql`
         avatar {
           children {
             ... on ImageSharp {
-              responsiveResolution(width: 80, height: 80, quality: 100) {
+              fixed(width: 80, height: 80, quality: 100) {
                 src
                 srcSet
               }
@@ -141,7 +146,7 @@ export const pageQuery = graphql`
       image {
         children {
           ... on ImageSharp {
-            responsiveResolution(width: 900, height: 300, quality: 100) {
+            fixed(width: 900, height: 300, quality: 100) {
               src
               srcSet
             }
@@ -170,7 +175,7 @@ export const pageQuery = graphql`
           image {
             children {
               ... on ImageSharp {
-                responsiveResolution(width: 300, height: 100) {
+                fixed(width: 300, height: 100) {
                   src
                   srcSet
                 }
@@ -182,7 +187,7 @@ export const pageQuery = graphql`
             avatar {
               children {
                 ... on ImageSharp {
-                  responsiveResolution(width: 36, height: 36) {
+                  fixed(width: 36, height: 36) {
                     src
                     srcSet
                   }
